+++ frontend/package.json
@@ -18,6 +18,7 @@
     "clsx": "^2.1.1",
     "node": "^22.6.0",
     "react": "^18.3.1",
+    "react-calendar": "^5.0.0",
     "react-dom": "^18.3.1",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7"
