+++ frontend/src/components/pages/DayComponent.tsx
